+++ ElvUI_LuckyoneUI/Init.lua
@@ -8,7 +8,7 @@ local format = format
 local _G = _G
 local GetAddOnMetadata = GetAddOnMetadata
 
-local L1UI = E:NewModule(addon, 'AceConsole-3.0', 'AceHook-3.0', 'AceEvent-3.0', 'AceTimer-3.0');
+local L1UI = E:NewModule(addon, 'AceConsole-3.0', 'AceHook-3.0', 'AceEvent-3.0', 'AceTimer-3.0')
 
 Engine[1] = L1UI -- LuckyoneUI
 Engine[2] = E -- ElvUI Engine
@@ -16,7 +16,7 @@ Engine[3] = L -- ElvUI Locales
 Engine[4] = V -- ElvUI PrivateDB
 Engine[5] = P -- ElvUI ProfileDB
 Engine[6] = G -- ElvUI GlobalDB
-_G[addon] = Engine;
+_G[addon] = Engine
 
 -- My characters
 local name = format('%s-%s', E.myname, E:ShortenRealm(E.myrealm))
@@ -43,9 +43,10 @@ local toons = {
 -- Constants
 L1UI.Config = {}
 L1UI.CreditsList = {}
+L1UI.Logo = 'Interface\\AddOns\\ElvUI_LuckyoneUI\\Media\\Textures\\Clover.tga'
 L1UI.Me = toons[name]
 L1UI.Name = '|cff4beb2cLuckyoneUI|r'
-L1UI.RequiredVersion = E.Retail and 12.75 or E.TBC and 2.42 or E.Classic and 1.67
+L1UI.RequiredVersion = E.Retail and 12.76 or E.TBC and 2.43 or E.Classic and 1.68
 L1UI.Version = GetAddOnMetadata(addon, 'Version')
 
 function L1UI:PLAYER_ENTERING_WORLD(_, initLogin, isReload)